账本切换

+++ app/src/main/java/com/cirnoteam/accountingassistant/activity/Status.java
@@ -9,6 +9,6 @@ import android.support.v7.app.AppCompatActivity;
 
 public class Status extends AppCompatActivity {
     public static String username;
-    public static int bookid;
+    public static Long bookid;
     public static int month = 0;
 }
